--- src/VirtualClient/VirtualClient.Actions.UnitTests/DCGMI/DCGMIHealth/DCGMIHealthCommandParserTests.cs
@@ -1,45 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-namespace VirtualClient.Actions
-{
-    using NUnit.Framework;
-    using System;
-    using System.Collections.Generic;
-    using System.IO;
-    using System.Linq;
-    using System.Reflection;
-    using System.Text;
-    using System.Threading.Tasks;
-    using VirtualClient.Contracts;
-
-    [TestFixture]
-    [Category("Unit")]
-    public class DCGMIHealthCommandParserTests
-    {   
-        [Test]
-        public void DCGMIHealthCommandParseParsesMetricsCorrectly()
-        {
-            string workingDirectory = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-            string outputPath = Path.Combine(workingDirectory, "Examples", "DCGMI", "DCGMIHealthCheckResults.json");
-            string rawText = File.ReadAllText(outputPath);
-
-            DCGMIHealthCommandParser testParser = new DCGMIHealthCommandParser(rawText);
-            IList<Metric> metrics = testParser.Parse();
-
-            Assert.AreEqual(1, metrics.Count);
-            MetricAssert.Exists(metrics, "Health Monitor Report_overallHealthValue", 1);
-        }
-
-        [Test]
-        public void DCGMIHealthCommandParseThrowsExceptionForIncorrectMetrics()
-        {
-            string workingDirectory = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-            string IncorrectoutputPath = Path.Combine(workingDirectory, "Examples", "DCGMI", "DCGMIIncorrectresults.json");
-            string rawText = File.ReadAllText(IncorrectoutputPath);
-            DCGMIHealthCommandParser testParser = new DCGMIHealthCommandParser(rawText);
-            SchemaException exception = Assert.Throws<SchemaException>(() => testParser.Parse());
-            StringAssert.Contains("The DCGMI Health output file has incorrect format for parsing", exception.Message);
-        }
-    }
-}
