--- src/VirtualClient/VirtualClient.Actions/DCGMI/DCGMIHealth/DCGMIHealthCommandParser.cs
@@ -1,80 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-namespace VirtualClient.Actions
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Text.Json;
-    using System.Threading.Tasks;
-    using Microsoft.CodeAnalysis;
-    using Newtonsoft.Json;
-    using VirtualClient.Contracts;
-
-    /// <summary>
-    /// Parser for DCGMIHealth command output document.
-    /// </summary>
-    public class DCGMIHealthCommandParser : MetricsParser
-    {
-        /// <summary>
-        /// Constructor for <see cref="DCGMIHealthCommandParser"/>
-        /// </summary>
-        /// <param name="rawText">Raw text to parse.</param>
-        public DCGMIHealthCommandParser(string rawText)
-            : base(rawText)
-        {
-        }
-
-        private List<Metric> Metrics { get; set; }
-
-        /// <inheritdoc/>
-        public override IList<Metric> Parse()
-        {
-            this.Preprocess();
-            this.Metrics = new List<Metric>();
-
-            double metricValue = 0;
-
-            var jsonDocument = JsonDocument.Parse(this.PreprocessedText);
-            try
-            {
-                // Get the "Overall Health" value
-                string overallHealthValue = jsonDocument.RootElement
-                    .GetProperty("body")
-                    .GetProperty("Overall Health")
-                    .GetProperty("value")
-                    .GetString();
-
-                // Get the "Health Monitor Report" header
-                string headerValue = jsonDocument.RootElement
-                    .GetProperty("header")
-                    .EnumerateArray()
-                    .First()
-                    .GetString();
-                if (overallHealthValue == "Healthy")
-                {
-                    metricValue = 1;
-                }
-                else
-                {
-                    metricValue = 0;
-                }
-
-                Dictionary<string, IConvertible> metadata = new Dictionary<string, IConvertible>()
-                         {
-                             { "overallHealthValue", overallHealthValue },
-                             { "headerValue", headerValue }
-                         };
-                this.Metrics.Add(new Metric(headerValue + "_" + "overallHealthValue", metricValue, metadata: metadata));
-            }
-            catch
-            {
-                throw new SchemaException("The DCGMI Health output file has incorrect format for parsing");
-            }
-
-            return this.Metrics;
-        }
-    }  
-}
